==> 0_helper_functions.R <==
library(terra)
library(rasterdiv)
library(GLCMTextures)
library(stringr)

### coefficient of variation:
source("./f_covar.R")

### Quick NDVI:
source("./f_ndvi.R")

### GLCM function:
source("./f_calc_spat_metrics.R")

### Get y variable:
source("./f_get_y_var.R")


### Linear models:
source("./f_linear_modeling.R")
